Backend. Custom exception added
https://github.com/Visual-Regression-Tracker/Visual-Regression-Tracker/issues/128
